re-introduce eager service loading
