+++ localstack-core/localstack/utils/bootstrap.py
@@ -332,13 +332,15 @@ def get_preloaded_services() -> Set[str]:
 
     The result is cached, so it's safe to call. Clear the cache with get_preloaded_services.cache_clear().
     """
-    services_env = os.environ.get("SERVICES", "").strip()
-    services = None
+    if not is_env_true("EAGER_SERVICE_LOADING"):
+        return set()
 
-    if services_env and is_env_true("EAGER_SERVICE_LOADING"):
+    services_env = os.environ.get("SERVICES", "").strip()
+    services = []
+
+    if services_env:
         # SERVICES and EAGER_SERVICE_LOADING are set
         # SERVICES env var might contain ports, but we do not support these anymore
-        services = []
         for service_port in re.split(r"\s*,\s*", services_env):
             # Only extract the service name, discard the port
             parts = re.split(r"[:=]", service_port)
@@ -346,24 +348,9 @@ def get_preloaded_services() -> Set[str]:
             services.append(service)
 
     if not services:
-        from localstack.services.plugins import SERVICE_PLUGINS
-
-        services = SERVICE_PLUGINS.list_available()
+        LOG.warning("No services set in SERVICES environment variable, skipping eager loading.")
 
     return resolve_apis(services)
-
-
-def should_eager_load_api(api: str) -> bool:
-    apis = get_preloaded_services()
-
-    if api in apis:
-        return True
-
-    for enabled_api in apis:
-        if api.startswith(f"{enabled_api}:"):
-            return True
-
-    return False
 
 
 def start_infra_locally():
